Times.Never is not properly evaluated
The behind loggerMock.Verify was called twice
for the times overload.

So if times is Times.Never and the sut implementation
is valid, the first verify call is not throwing, as expected.

But the second loggerMock.Verify was also called with default values
and the test reported a false negative, as the sut implementation
was correct, but test improperly failed expecting a call on sut's method.

Fixes issue #4